Small bug fixes
1. In the short delay between hiding the message, updating its
contents, and reshowing it, a user could interact with the UI,
and could potentially accidentaly start another download/search
whilst the first is still ongoing. This fix ensures the UI is
uninteractable (aside from the cancel button of course), until
the operation is finished or canceled.

2. The status of the message box's visible status wasn't being
updated when it was closed by the user, using the Ok/Cancel
buttons. This caused incorrect messages to appear. This fix
causes the visible status to be correctly updated.