Better http code checks and (en|de)coding of urls
We now check that we get a 200 success code before downloading.
This prevents us downloading invalid images, for example
if the server has an internal 500 code, the image will be written
with this text rather than skipping the image.

We now also use the nice Network.URI tools to encode and decode
our urls, which fixes a few bugs which have been experienced with
odd characters such as Ã

Furthermore, we also now handle some odd urls which don't appear
to return the expected page order urls in order. This prevents
a previous infinite/very long loop on certain inputs.